requirements.txt
botbuilder-integration-aiohttp>=4.11.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
datatypes-date-time>=1.0.0.a2



botbuilder-core==4.13.0
botframework-connector==4.13.0
botbuilder-schema==4.13.0
botbuilder-dialogs==4.13.0
botbuilder-ai==4.13.0 #install luis 0.2.0 instead of 0.7.0
botbuilder-applicationinsights==4.13.0
botbuilder-integration-applicationinsights-aiohttp==4.13.0
azure-cognitiveservices-language-luis==0.2.0
msrest==0.6.10
datatypes-date-time>=1.0.0.a1
aiohttp==3.6.2
requests==2.23.0
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
Django==2.2.6
djangorestframework==3.10.3
emoji==1.6.3
datatypes-date-time==1.0.0a2
Flask==1.1.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
uvicorn==0.17.6
pytest